xps backup

+++ controllerFirmware/src/fonts/font.h
@@ -5,10 +5,10 @@
 
 class Font {
   private:
-    int charWidth, charHeight, bitWidth, nchars;
-    const uint16_t* bitmap;
+    unsigned short charWidth, charHeight, bitWidth, nchars;
+    const unsigned short* bitmap;
   public:
-    Font(unsigned int charWidth, unsigned int charHeight, const uint16_t* bitmap) {
+    Font(unsigned int charWidth, unsigned int charHeight, const unsigned short* bitmap) {
       this->charWidth = charWidth;
       this->charHeight = charHeight;
       this->bitWidth = charWidth + (8-charWidth%8);
@@ -20,11 +20,11 @@ class Font {
       return pgm_read_word(bitmap + character*(this->charHeight+1) + y+1) & (0x8000>>x);
     }
 
-    int getCharWidth(char character) {
-      return pgm_read_word(bitmap + character*(this->charHeight+1));
+    unsigned short getCharWidth(char character) {
+      return pgm_read_word(bitmap + (unsigned short)character*(this->charHeight+1));
     }
 
-    int getCharHeight() {
+    unsigned short getCharHeight() {
       return charHeight;
     }
 };
